CDAP-19706 used Mui Menu and deleted Custom Menu
